refactor: FakeCars 클래스 public으로 변경하고 fake 패키지에 넣음
